fix(consent): release Hydrogen's analytics queue in non-banner regions
Root cause found on 7Diamonds after the earlier fixes didn't resolve it, and it
affects every storefront with `withPrivacyBanner: true`.

Hydrogen's own <ShopifyAnalytics> registers a consumer we don't control,
"Internal_Shopify_Analytics", and only reports it ready once its internal
`privacyReady` flips. That has exactly two sources:

  onReady: () => !consent.withPrivacyBanner && setPrivacyReady(true)
  onVisitorConsentCollected: (c) => { ...; setPrivacyReady(true); }

`withPrivacyBanner: true` disables the first. The second is the event Shopify
never dispatches where no cookie banner is shown. So the register never readies,
`areRegistersReady()` is permanently false, and Hydrogen parks every analytics
event in `waitForReadyQueue` rather than publishing it.

The symptom is maximally misleading: vendor libraries load, subscriptions report
ready, `analyticsProcessingAllowed()` is true, `shop` resolves — and nothing
publishes. Shopify's own analytics is equally dead, which is the tell.

Verified present in both @shopify/hydrogen 2025.7.3 and 2026.1.0, so this is not
a version regression — it's inherent to the `withPrivacyBanner: true` +
no-banner-region combination.

Folds the previous `useGlobalPrivacyControl` into `useConsentBridge`, which
commits the resolved consent once — releasing Hydrogen's queue — and applies the
GPC override to `sale_of_data` in the same write rather than issuing two.
Skipped while a banner is shown, since there the shopper's own interaction fires
the event.

CONSENT.md documents the deadlock, its confusing symptom set, and that
`withPrivacyBanner: false` (or omitted) is immune.

Refs PLT-4029
